requirements.txt
Bottleneck==1.3.4
certifi==2022.5.18.1
charset-normalizer==2.0.4
click==8.1.3
colorama==0.4.4
deprecation==2.1.0
docopt==0.6.2
elastic-transport==8.1.2
elasticsearch==7.17.4
elasticsearch-dsl==7.4.0
Flask==2.1.2
future==0.17.1
idna==3.2
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.0.1
MarkupSafe==2.1.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
nltk==3.7
num2words==0.5.10
numexpr==2.8.1
numpy==1.21.5
packaging==21.3
pandas==1.3.5
pip==21.0.1
pyaml==19.12.0
pyparsing==3.0.4
python-crfsuite==0.9.7
python-dateutil==2.8.2
pytz==2022.1
PyYAML==5.4.1
regex==2022.6.2
requests==2.26.0
scikit-learn==0.22.2.post1
scipy==1.7.1
setuptools==52.0.0.post20210125
six==1.16.0
sklearn-crfsuite==0.3.6
snips-nlu==0.20.2
snips-nlu-city-en==0.2.0
snips-nlu-country-en==0.2.0
snips-nlu-en==0.2.3
snips-nlu-parsers==0.4.3
snips-nlu-utils==0.9.1
tabulate==0.8.9
textblob==0.17.1
tqdm==4.62.1
typing-extensions==4.2.0
urllib3==1.26.6
Werkzeug==2.1.2
wheel==0.37.0
wincertstore==0.2
zipp==3.8.0
